+++ test_autolens/interferometer/model/test_result_interferometer.py
@@ -1,0 +1,9 @@
+import os
+
+
+directory = os.path.dirname(os.path.realpath(__file__))
+
+
+class TestResultInterferometer:
+
+    pass
